requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.6.0
spacy==3.7.6
python-dotenv==1.0.1
pydantic==2.8.2
ollama==0.3.0
